% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bear.R
\docType{data}
\name{bear}
\alias{bear}
\title{Bearings data set.}
\format{
A data frame with 15 rows and 3 variables:
\describe{
  \item{DAY}{length in cm}
  \item{DEFECTS}{Dollars damage}
  \item{PCTDEF}{Garbage}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
bear
}
\description{
A data set from MS
}
\keyword{datasets}
