% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullseye.R
\name{bullseye}
\alias{bullseye}
\title{Bull's eye function}
\usage{
bullseye(n = 30, iter = 10000, mu = 10, sigma = 10, nadj = 1)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{mu}{Population mean}

\item{sigma}{Population standard deviation}

\item{nadj}{Adjustment to `n-nadj` 1 or 0}
}
\value{
Plots a bull's eye target
}
\description{
This function shows how unbiased estimators work
}
\examples{
bullseye()
}
