% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{logbin2}
\alias{logbin2}
\title{logbin2}
\usage{
logbin2(theta)
}
\arguments{
\item{theta}{Probability parameter}
}
\value{
Log-likelihood value
}
\description{
Log-likelihood for a binomial combination.
}
