% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmat.R
\name{mvmat}
\alias{mvmat}
\title{Multivariate Normal Distribution}
\usage{
mvmat(sig11 = 100, sig22 = 400, sig12 = 100, rangxy = 15)
}
\arguments{
\item{sig11}{Population variance 11}

\item{sig22}{Population variance 22}

\item{sig12}{Population covariance 12}

\item{rangxy}{How many std dev to make the range from}
}
\value{
invisible list containing the centered data
}
\description{
Make a zero centered sample
}
\examples{
mvmat()
}
