% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{A function that bootstraps}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{nu. of iterations}

\item{x}{vector of originating sample}

\item{fun}{function that creates the statistic}

\item{alpha}{error rate}

\item{cx}{character expansion}

\item{...}{further parameters to be sent to the histogram}
}
\value{
vector of simulations and a histogram
}
\description{
This is the main engine for bootstrapping
}
\examples{
myboot2(x=rnorm(30))
}
