% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myciforl.R
\name{myCIforl}
\alias{myCIforl}
\title{Confidence interval for l=a'beta}
\usage{
myCIforl(alpha, betahat, X, a, s)
}
\arguments{
\item{alpha}{confidence level}

\item{betahat}{point estimates for beta, where Y=beta_0+X_1*beta_1+...}

\item{X}{design matrix nx(k+1) where the first column is all 1s (for beta_0)}

\item{a}{vector with linear combination terms for beta}

\item{s}{estimate of the standard deviation}
}
\value{
confidence interval for l
}
\description{
This function gives confidence intervals for a multiple regression model, where l=a'*beta. Meaning l is any linear combination of beta parameters, where
Y=beta_0+beta_1*x_1+...
}
\examples{
ci=myCIforl(alpha,betahat,X,a,s)
}
