% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{CLT function}
\usage{
myclt(n, iter, a = 0, b = 1)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations (bigger the better - choice depends on computing power and how long you wish to wait)}

\item{a}{first parameter of uniform}

\item{b}{second parameter of uniform}
}
\value{
Histogram and a vector of stats
}
\description{
This function creates a samlpling distribution of a statistic
}
\examples{
myclt(n=10,iter=10000)
}
