% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{An example of a Function to be zeroed}
\usage{
myfun(x)
}
\arguments{
\item{x}{Parameter values for which a zero is captured}
}
\value{
y value f(x)
}
\description{
An example of a Function to be zeroed
}
\examples{
myfun(10)
}
