% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myl.R
\name{myl}
\alias{myl}
\title{My ci for E(l)}
\usage{
myl(Y, X, alpha, a, l)
}
\arguments{
\item{Y}{dependent data}

\item{X}{Design Matrix}

\item{alpha}{confidence level}

\item{a}{l=a'B}

\item{l}{l=a'B}
}
\value{
ci, scatterplots of data
}
\description{
This Function creates confidence intervals for any linear combination of betas
}
\examples{
ci=myl(Y,X,alpha,a,l)
}
