% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{Log-likelihood function (default = "logbin2")}

\item{theta}{Numeric vector of theta values}
}
\value{
Theta value at the maximum likelihood
}
\description{
Computes and plots likelihood and finds the MLE.
}
