% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{A Normal curve with area}
\usage{
myncurve(x, mu, sigma)
}
\arguments{
\item{x}{The upper bound of the area}

\item{mu}{the mean}

\item{sigma}{the standard deviation}
}
\value{
A curve, shaded region and area
}
\description{
A Normal curve with area
}
\examples{
myncurve(x=5,mu =10, sigma =4)
}
