% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynewt.R
\name{mynewt}
\alias{mynewt}
\title{Newton Raphson Algorithm}
\usage{
mynewt(x0, f = myfun, delta = 1e-12, epsilon = 1e-12, th1 = c(0, 2
  * pi), th2 = c(0, pi/2), parameter = expression(theta))
}
\arguments{
\item{x0}{Initial value}

\item{f}{function to find the roots of}

\item{delta}{Used to find the derivative , h}

\item{epsilon}{Accuracy}

\item{th1}{Domain for Upper plot}

\item{th2}{Domain for lower (magnified) plot}

\item{parameter}{Name of the parameter to be estimated}
}
\value{
invisible list of root ,x,y
}
\description{
Find roots calculating estimates of the slope internally
}
\examples{
root=mynewt(1.0,th2=c(1,1.4))$root
s12=100
s11=200
s22=400
#Test the answer in two ways
myfun(root)
}
